Replaced MPT with Verkle Trie

+++ libs/bullet_db/src/state/ext.cpp
@@ -1,13 +1,7 @@
-#include "blake3.h"
-#include "nodes.h"
-#include <cassert>
-#include <cstdint>
-#include <cstdio>
-#include <optional>
-#include <tuple>
-#include <vector>
+#include "verkle.h"
 
 Extension::Extension(std::optional<NodeId> id, std::optional<ByteSlice*> buffer) : path_(32){
+    commit_ = new_p1();
     if (!id.has_value()) return;
     id_ = id.value();
     child_id_ = u64_to_array(u64_from_array(id_) * ORDER);
@@ -15,13 +9,9 @@ Extension::Extension(std::optional<NodeId> id, std::optional<ByteSlice*> buffer)
     if (!buffer.has_value()) return;
     ByteSlice* buff = buffer.value();
 
-    auto cursor = buff->begin() + 1;
-    std::ranges::copy(cursor, cursor+32, hash_.begin());
-    cursor += 32;
+    commit_from_bytes(buff->subspan(1,49).data(), commit_);
 
-    std::ranges::copy(cursor, cursor+32, child_hash_.begin());
-    cursor += 32;
-
+    auto cursor = buff->begin() + 49;
 
     uint64_array path_len;
     std::ranges::copy(cursor, cursor+8, path_len.begin());
@@ -35,10 +25,7 @@ Extension::Extension(std::optional<NodeId> id, std::optional<ByteSlice*> buffer)
 }
 
 NodeId* Extension::get_child_id() { return &child_id_; }
-Hash* Extension::get_child_hash() { return &child_hash_; }
-void Extension::set_child(Hash &hash) {
-    child_hash_ = hash;
-}
+void Extension::set_child(Commitment &c) { commit_ = c; }
 
 Path* Extension::get_path() { return &path_; }
 void Extension::set_path(ByteSlice path) {
@@ -90,22 +77,8 @@ void Extension::change_id(const NodeId &id, Ledger &ledger) {
     }
 }
 
-Hash Extension::derive_hash() {
-    blake3_hasher hasher;
-    blake3_hasher_init(&hasher);
-    blake3_hasher_update(
-        &hasher, 
-        child_hash_.data(),
-        child_hash_.size());
-
-    Hash hash;
-    blake3_hasher_finalize(
-        &hasher, 
-        reinterpret_cast<uint8_t*>(hash.data()), 
-        hash.size());
-
-    hash_ = hash;
-    return hash;
+Commitment* Extension::derive_commitment(Ledger &ledger) { 
+    return &commit_; 
 }
 
 std::vector<byte> Extension::to_bytes() {
@@ -113,8 +86,8 @@ std::vector<byte> Extension::to_bytes() {
     buffer.reserve(EXT_SIZE);
 
     buffer.push_back(EXT);
-    buffer.insert(buffer.end(), hash_.begin(), hash_.end());
-    buffer.insert(buffer.end(), child_hash_.begin(), child_hash_.end());
+    auto commit_bytes = compress_p1(commit_);
+    buffer.insert(buffer.end(), commit_bytes.begin(), commit_bytes.end());
 
     uint64_array path_len = u64_to_array(path_.size());
     buffer.insert(buffer.end(), path_len.begin(), path_len.end());
@@ -127,31 +100,28 @@ std::vector<byte> Extension::to_bytes() {
     return buffer;
 }
 
-optional<Hash> Extension::search( 
+optional<Commitment> Extension::search( 
     Ledger &ledger, 
-    ByteSlice &nibbles
+    ByteSlice nibbles
 ) {
     if (NodeId* next = get_next_id(nibbles)) {
         if (Node* n = ledger.load_node(*next)) {
-            auto remaining_nibbles = nibbles.subspan(path_.size());
-            return n->search(ledger, remaining_nibbles);
+            return n->search(ledger, nibbles.subspan(path_.size()));
         }
     }
     return std::nullopt;
 }
 
-optional<Hash> Extension::virtual_put(
+optional<Commitment> Extension::virtual_put(
     Ledger &ledger, 
-    ByteSlice &nibbles,
+    ByteSlice nibbles,
     const ByteSlice &key, 
-    const Hash &val_hash
+    const Commitment &val_commitment
 ) {
     if (NodeId* next = get_next_id(nibbles)) {
         if (Node* n = ledger.load_node(*next)) {
-            auto remaining_nibbles = nibbles.subspan(path_.size());
-            auto res = n->virtual_put(ledger, remaining_nibbles, key, val_hash);
-            if (res.has_value())
-                return derive_value_hash(res.value().data(), 32);
+            auto remaining = nibbles.subspan(path_.size());
+            return n->virtual_put( ledger, remaining, key, val_commitment);
         }
         return std::nullopt;
     }
@@ -166,60 +136,37 @@ optional<Hash> Extension::virtual_put(
     auto prefix = nibbles.first(shared_path);
     nibbles = nibbles.subspan(shared_path);
 
-    Hash leaf_hash = derive_leaf_hash(key, val_hash);
-
-    branch.insert_child(nibbles[0], leaf_hash);
+    branch.insert_child(nibbles[0], val_commitment);
 
     // if self.path shares nothing with new key
     // remove old self id
     // connect new_branch to self_copy using popped nibble from self
     // return new_branch to parent
     if (shared_path == 0) {
+
         byte nib = path_.pop_front().value();
-
-        // handle edge case where pathlen becomes 0.
-        if (path_.empty()) {
-            branch.insert_child(nib, child_hash_);
-        } else {
-            branch.insert_child(nib, hash_);
-        }
-
+        branch.insert_child(nib, commit_);
         path_.push_front(nib);
-        // parent will point to new_branch instead of self
-        return branch.derive_hash();
-    }
 
-    // if there is path remaining use it to create new extension
-    // new extension will connect new_branch to self.child
-    // else point branch to self.child directly.
+        // parent will point to new_branch instead of self
+        return *branch.derive_commitment(ledger);
+    }
 
     byte nib = path_.get(shared_path).value();
-    if (path_.size() > shared_path) {
-        // new branch points to new extension
-        branch.insert_child(nib, hash_);
-    } else {
-        // point to child
-        branch.insert_child(nib, child_hash_);
-    }
-
-    // derive final virtual state
-    return derive_value_hash(child_hash_.data(), 32);
+    branch.insert_child(nib, commit_);
+    return commit_;
 }
 
-optional<Hash> Extension::put(
+optional<Commitment*> Extension::put(
     Ledger &ledger, 
-    ByteSlice &nibbles, 
+    ByteSlice nibbles, 
     const ByteSlice &key, 
-    const Hash &val_hash
+    const Commitment &val_commitment
 ) {
     if (NodeId* next = get_next_id(nibbles)) {
         if (Node* n = ledger.load_node(*next)) {
-
-            std::span<byte> remaining_nibbles = nibbles.subspan(path_.size());
-
-            optional<Hash> hash = n->put(ledger, remaining_nibbles, key, val_hash);
-
-            if (hash.has_value()) return derive_value_hash(hash.value().data(), 32);
+            auto remaining_nibbles = nibbles.subspan(path_.size());
+            return n->put(ledger, remaining_nibbles, key, val_commitment);
         }
         return std::nullopt;
     }
@@ -256,11 +203,10 @@ optional<Hash> Extension::put(
     uint64_t new_branch_child_id = (branch_id * ORDER) + idx;
 
     Leaf* leaf = ledger.new_cached_leaf(new_branch_child_id);
-    leaf->set_value_hash(val_hash);
+    leaf->set_commitment(val_commitment);
     leaf->set_path(nibbles.subspan(1));
 
-    Hash leaf_hash = leaf->derive_real_hash(key);
-    branch->insert_child(nibbles[0], leaf_hash);
+    branch->insert_child(nibbles[0], val_commitment);
 
     // if self.path shares nothing with new key
     // remove old self id
@@ -276,7 +222,7 @@ optional<Hash> Extension::put(
             new_child_id = (branch_id * ORDER) + nib_num;
             uint64_array new_id_raw = u64_to_array(new_child_id);
             child->change_id(new_id_raw, ledger);
-            branch->insert_child(nib, child_hash_);
+            branch->insert_child(nib, commit_);
             
         } else {
             uint64_t new_id = (branch_id * ORDER) + nib_num;
@@ -290,8 +236,7 @@ optional<Hash> Extension::put(
             // recover self change_id and reinsert into cache
             uint64_array new_raw_id = u64_to_array(new_id);
             self.value()->change_id(new_raw_id, ledger);
-            Hash self_hash = self.value()->derive_hash();
-            branch->insert_child(nib, self_hash);
+            branch->insert_child(nib, commit_);
 
             ledger.cache_node(std::move(self.value()));
         }
@@ -299,7 +244,7 @@ optional<Hash> Extension::put(
         ledger.cache_node(std::move(child));
 
         // parent will point to new_branch instead of self
-        return branch->derive_hash();
+        return branch->derive_commitment(ledger);
     }
 
     // if there is path remaining use it to create new extension
@@ -331,11 +276,10 @@ optional<Hash> Extension::put(
         child->change_id(new_id, ledger);
 
         // set new child hash
-        new_ext->set_child(*child->get_hash());
+        new_ext->set_child(*child->get_commitment());
 
         // new branch points to new extension
-        Hash ext_hash = new_ext->derive_hash();
-        branch->insert_child(nib, ext_hash);
+        branch->insert_child(nib, *new_ext->get_commitment());
 
     } else {
         // derive and set new child id
@@ -344,43 +288,36 @@ optional<Hash> Extension::put(
         child->change_id(new_id, ledger);
 
         // point branch to self.child
-        Hash child_hash = child->derive_hash();
-        branch->insert_child(nib, child_hash);
+        branch->insert_child(nib, *child->get_commitment());
     }
 
     // have to recache old_child
     ledger.cache_node(std::move(child));
 
     // set self to point to branch return self
-    Hash branch_hash = branch->derive_hash();
-    set_child(branch_hash);
-    return derive_hash();
+    set_child(*branch->derive_commitment(ledger));
+    return get_commitment();
 }
 
-optional<std::tuple<Hash, ByteSlice>> Extension::remove(
+optional<std::tuple<Commitment, bool, ByteSlice>> Extension::remove(
     Ledger &ledger, 
-    ByteSlice &nibbles
+    ByteSlice nibbles
 ) {
     if (NodeId* next = get_next_id(nibbles)) {
         if (Node* n = ledger.load_node(*next)) {
             std::span<byte> remaining_nibbles = nibbles.subspan(path_.size());
 
             auto response = n->remove(ledger, remaining_nibbles);
-            if (!response.has_value()) return std::nullopt;
+            if (response.has_value()) {
+                auto [new_commit, removed, path_ext] = response.value();
 
-            auto [hash, path_ext] = response.value();
+                if (path_ext.size() > 0) 
+                    for (auto nib: path_ext) path_.push_back(nib);
 
-            if (path_ext.size() > 0) 
-                for (auto nib: path_ext) 
-                    path_.push_back(nib);
-
-            if (is_zero(hash)) {
-                ledger.delete_node(id_);
-                return std::make_tuple(hash, path_ext);
+                if (removed) ledger.delete_node(id_);
+                else set_child(new_commit);
             }
-
-            set_child(hash);
-            return std::make_tuple(derive_hash(), ByteSlice{});
+            return response;
         }
     }
     return std::nullopt;
